requirements.txt
Flask
google-auth
google-auth-oauthlib
google-auth-httplib2
google-ads
requests
google-api-python-client
python-dotenv
